+++ demo/src/app/single/single.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {ShareButton, ShareProvider} from "ng2-sharebuttons";
+import {ShareButton, ShareProvider} from "../ng2-sharebuttons";
 
 @Component({
     selector: 'app-single',